requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
python-decouple>=3.8,<4.0
gunicorn>=21.2,<22.0